Disabled buttons when no game is selected, watermark to indicate that a game should be chosen. Separation of components
