+++ src/App.module.scss
@@ -42,20 +42,4 @@
     display: flex;
     flex-flow: column nowrap;
   }
-
-  .newVersion {
-    :global(.text) {
-      background-color: #ffffcc;
-    }
-  }
-}
-
-.github {
-  width: 48px;
-  height: 48px;
-  min-height: 48px !important;
-  min-width: 48px !important;
-  font-size: 32px !important;
-  padding: 0 !important;
-  align-self: center;
 }
