+++ src/components/Options/Options.tsx
@@ -9,6 +9,7 @@ const Options: React.FC = React.memo(() => {
   const search = useStore(useCallback((state) => state.search, []));
   const resetAll = useStore(useCallback((state) => state.resetAll, []));
   const addEncounter = useStore(useCallback((state) => state.addEncounter, []));
+  const selectedGame = useStore(useCallback((state) => state.selectedGame, []));
   const darkMode = useStore(useCallback((state) => state.darkMode, []));
   const [searchText, setSearchText] = useState(text);
   const debouncedValue = useDebounce<string>(searchText, 250);
@@ -45,6 +46,7 @@ const Options: React.FC = React.memo(() => {
       <Input
         aria-label="search"
         className={styles.search}
+        disabled={!selectedGame}
         fluid
         icon
         id="search-filter"
@@ -62,7 +64,12 @@ const Options: React.FC = React.memo(() => {
           closeOnDimmerClick
           open={open}
           trigger={
-            <Button color="green" inverted={darkMode} onClick={() => setOpen(true)}>
+            <Button
+              color="green"
+              disabled={!selectedGame}
+              inverted={darkMode}
+              onClick={() => setOpen(true)}
+            >
               ADD ENCOUNTER
               <i className="icon plus" />
             </Button>
@@ -80,7 +87,12 @@ const Options: React.FC = React.memo(() => {
             </Button>
           </Modal.Actions>
         </Modal>
-        <Button color="red" inverted={darkMode} onClick={() => setConfirm(true)}>
+        <Button
+          color="red"
+          disabled={!selectedGame}
+          inverted={darkMode}
+          onClick={() => setConfirm(true)}
+        >
           RESET ALL
           <i className="icon close" />
         </Button>
