+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export { default as About } from './About/About';
 export { default as BadgeEditor } from './BadgeEditor/BadgeEditor';
 export { default as Badges } from './Badges/Badges';
 export { default as Contact } from './Contact/Contact';
@@ -5,5 +6,6 @@ export { default as Encounters } from './Encounters/Encounters';
 export { default as Nickname } from './Nickname/Nickname';
 export { default as Options } from './Options/Options';
 export { default as Pokemon } from './Pokemon/Pokemon';
+export { default as Share } from './Share/Share';
 export { default as Status } from './Status/Status';
 export { default as Tracker } from './Tracker/Tracker';
